--- system/cu/cu.h
@@ -1,92 +1,0 @@
-/****************************************************************************
- * apps/system/cu/cu.h
- *
- * SPDX-License-Identifier: BSD-3-Clause
- * SPDX-FileCopyrightText: 2014 sysmocom - s.f.m.c. GmbH. All rights reserved.
- * SPDX-FileContributor: Harald Welte <hwelte@sysmocom.de>
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
- * 3. Neither the name NuttX nor the names of its contributors may be
- *    used to endorse or promote products derived from this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
- * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *
- ****************************************************************************/
-
-#ifndef __APPS_SYSTEM_CU_CU_H
-#define __APPS_SYSTEM_CU_CU_H
-
-/****************************************************************************
- * Included Files
- ****************************************************************************/
-
-#include <nuttx/config.h>
-
-#include <stdint.h>
-#include <stdbool.h>
-#include <semaphore.h>
-#include <termios.h>
-
-/****************************************************************************
- * Pre-processor Definitions
- ****************************************************************************/
-
-/* Configuration ************************************************************/
-
-#ifndef CONFIG_SYSTEM_CUTERM_DEFAULT_DEVICE
-#  define CONFIG_SYSTEM_CUTERM_DEFAULT_DEVICE "/dev/ttyS0"
-#endif
-
-#ifndef CONFIG_SYSTEM_CUTERM_DEFAULT_BAUD
-#  define CONFIG_SYSTEM_CUTERM_DEFAULT_BAUD 115200
-#endif
-
-/****************************************************************************
- * Public Types
- ****************************************************************************/
-
-/* All terminal state data is packaged in a single structure to minimize
- * name conflicts with other global symbols -- a poor man's name space.
- */
-
-struct cu_globals_s
-{
-  int devfd;             /* I/O data to serial port */
-  int stdfd;             /* I/O data to standard console */
-  int escape;            /* Escape char */
-  struct termios devtio; /* Original serial port setting */
-  struct termios stdtio; /* Original standard console setting */
-  pthread_t listener;    /* Terminal listener thread */
-  bool force_exit;       /* Force exit */
-};
-
-/****************************************************************************
- * Public Data
- ****************************************************************************/
-
-/****************************************************************************
- * Public Function Prototypes
- ****************************************************************************/
-
-#endif /* __APPS_SYSTEM_CU_CU_H */
